+++ tick_utils.h
@@ -15,15 +15,18 @@
 /****************************************************************/
 
 
+#if (__CORTEX_M != 0U)
 /*!\brief Microseconds delay generator (blocking)
 ** \note Please keep in mind Delay_us is blocking code execution for the given amount of time (except interrupts).
 ** \note On some Cortex cores, need to enable access to DWT prior to init Delay_us (automatically done only for M7 core).
+** \note Cortex cores M0/M0+ doesn't have DWT peripheral (the function is disabled). TODO: find another way to implement Delay_us for M0/M0+ cores
 ** \warning Delay initialization instructions being executed on first call, this delay will last a little longer.
-**			If delay really needs to be accurate on first go, call Delay_us(1) at startup before realy using it.
+**			If delay really needs to be accurate on first go, call Delay_us(1) at startup before really using it.
 ** \warning As for functions like HAL_Delay, don't expect accurate delay for just a few us (through tests, overtaking is observed weighted by System Clock)
 ** \param[in] delay - delay (in µs)
 **/
 void Delay_us(const uint32_t us);
+#endif
 
 
 /****************************************************************/